pyscf-front/backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyscf-front"
version = "0.1.0"
description = "Quantum chemistry calculation native application frontend"
authors = [
    {name = "PySCF_Front Team", email = "dev@pyscf-front.org"}
]
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.13"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Chemistry",
]

dependencies = [
    "pyscf>=2.9.0",
    "numpy>=1.26.4",
    "scipy>=1.13.1",
    "h5py>=3.11.0",
    "mysql-connector-python>=9.3.0",
    "sqlalchemy>=2.0.30",
    "alembic>=1.13.1",
    "grpcio>=1.73.0",
    "grpcio-tools>=1.73.0",
    "protobuf>=5.27.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.1",
    "rdkit>=2024.03.3",
    "pandas>=2.2.2",
    "matplotlib>=3.9.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.7.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.2",
    "pytest-asyncio>=0.23.7",
    "pytest-cov>=5.0.0",
    "black>=24.4.2",
    "isort>=5.13.2",
    "flake8>=7.0.0",
    "mypy>=1.10.0",
]

docs = [
    "sphinx>=7.3.7",
    "sphinx-rtd-theme>=2.0.0",
]

gpu = [
    "gpu4pyscf-cuda11x>=1.4.0",  # CUDA 11.x
    # "gpu4pyscf-cuda12x>=1.4.0",  # CUDA 12.x
]

mcp = [
    "anthropic-model-context-protocol>=0.1.0",
    "websockets>=12.0",
]

[project.urls]
Homepage = "https://github.com/pyscf-front/pyscf-front"
Repository = "https://github.com/pyscf-front/pyscf-front.git"
Documentation = "https://pyscf-front.readthedocs.io"
"Bug Tracker" = "https://github.com/pyscf-front/pyscf-front/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["pyscf_front"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--strict-markers --disable-warnings"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]